# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for registration.",
        "version": "1.0.0",
        "title": "Registration API"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
...
    "paths": {
...
        "/v1/internal/agents": {
            "post": {
                "tags": ["Agents"],
                "summary": "Create a new agent",
                "description": "Endpoint to create a new agent",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                {
                    "in": "body",
                    "name": "body",
                    "description": "Agent object that needs to be added",
                    "required": true,
                    "schema": {
                    "$ref": "#/definitions/Agent"
                    }
                }
                ],
                "responses": {
                "200": {
                    "description": "Agent created successfully"
                },
                "400": {
                    "description": "Invalid input"
                }
                }
            },
            "get": {
                "tags": ["Agents"],
                "summary": "Get all agents",
                "description": "Endpoint to get all agents",
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "description": "A list of agents",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Agent"
                            }
                        }
                    },
                    "500": {
                        "description": "Error retrieving agents"
                    }
                }
            }
        },
...
        "/v1/internal/affiliates": {
            "get": {
                "tags": ["Affiliates"],
                "summary": "Get all affiliates",
                "description": "Endpoint to get all affiliates",
                "produces": ["application/json"],
                "responses": {
                "200": {
                    "description": "A list of affiliates",
                    "schema": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Affiliate"
                    }
                    }
                },
                "500": {
                    "description": "Error retrieving affiliates"
                }
                }
            }
        },
        "/v1/internal/insurance-objects": {
            "post": {
                "tags": ["InsuranceObjects"],
                "summary": "Create a new insurance object",
                "description": "Endpoint to create a new insurance object",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                {
                    "in": "body",
                    "name": "body",
                    "description": "InsuranceObject object that needs to be added",
                    "required": true,
                    "schema": {
                    "$ref": "#/definitions/InsuranceObject"
                    }
                }
                ],
                "responses": {
                "201": {
                    "description": "InsuranceObject created successfully"
                },
                "400": {
                    "description": "Invalid input"
                }
                }
            },
            "get": {
                "tags": ["InsuranceObjects"],
                "summary": "Get all insurance objects",
                "description": "Endpoint to get all insurance objects",
                "produces": ["application/json"],
                "responses": {
                "200": {
                    "description": "A list of insurance objects",
                    "schema": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/InsuranceObject"
                    }
                    }
                },
                "500": {
                    "description": "Error retrieving insurance objects"
                }
                }
            }
        },
        "/v1/internal/insurance-risks": {
            "post": {
                "tags": ["InsuranceRisks"],
                "summary": "Create a new insurance risk",
                "description": "Endpoint to create a new insurance risk",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                {
                    "in": "body",
                    "name": "body",
                    "description": "InsuranceRisk object that needs to be added",
                    "required": true,
                    "schema": {
                    "$ref": "#/definitions/InsuranceRisk"
                    }
                }
                ],
                "responses": {
                "201": {
                    "description": "InsuranceRisk created successfully"
                },
                "400": {
                    "description": "Invalid input"
                }
                }
            },
            "get": {
                "tags": ["InsuranceRisks"],
                "summary": "Get all insurance risks",
                "description": "Endpoint to get all insurance risks",
                "produces": ["application/json"],
                "responses": {
                "200": {
                    "description": "A list of insurance risks",
                    "schema": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/InsuranceRisk"
                    }
                    }
                },
                "500": {
                    "description": "Error retrieving insurance risks"
                }
                }
            }
        },
...
        "/v1/public/contracts": {
            "post": {
                "tags": ["Contracts"],
                "summary": "Create a new contract",
                "description": "Endpoint to create a new contract",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                {
                    "in": "body",
                    "name": "body",
                    "description": "Contract object that needs to be added",
                    "required": true,
                    "schema": {
                    "$ref": "#/definitions/Contract"
                    }
                }
                ],
                "responses": {
                "201": {
                    "description": "Contract created successfully"
                },
                "400": {
                    "description": "Invalid input"
                }
                }
            },
            "get": {
                "tags": ["Contracts"],
                "summary": "Get all contracts",
                "description": "Endpoint to get all contracts",
                "produces": ["application/json"],
                "responses": {
                "200": {
                    "description": "A list of contracts",
                    "schema": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Contract"
                    }
                    }
                },
                "500": {
                    "description": "Error retrieving contracts"
                }
                }
            }
        },
...
    "/v1/public/policies": {
        "post": {
        "tags": ["Policies"],
        "summary": "Create a new policy",
        "description": "Endpoint to create a new policy",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
            {
            "in": "body",
            "name": "body",
            "description": "Policy object that needs to be added",
            "required": true,
            "schema": {
                "$ref": "#/definitions/Policy"
            }
            }
        ],
        "responses": {
            "201": {
            "description": "Policy created successfully"
            },
            "400": {
            "description": "Invalid input"
            }
        }
        },
        "get": {
        "tags": ["Policies"],
        "summary": "Get all policies",
        "description": "Endpoint to get all policies",
        "produces": ["application/json"],
        "responses": {
            "200": {
            "description": "A list of policies",
            "schema": {
                "type": "array",
                "items": {
                "$ref": "#/definitions/Policy"
                }
            }
            },
            "500": {
            "description": "Error retrieving policies"
            }
        }
        }
    }
    },
    "definitions": {
        "Policy": {
            "type": "object",
            "required": ["agent", "contract", "insuranceSum", "price", "startDate", "endDate"],
            "properties": {
                "agent": {
                    "type": "string",
                    "format": "ObjectId"
                },
                "contract": {
                    "type": "string",
                    "format": "ObjectId"
                },
                "insuranceSum": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "startDate": {
                    "type": "string",
                    "format": "date"
                },
                "endDate": {
                    "type": "string",
                    "format": "date"
                }
            }
        },
        "Contract": {
            "type": "object",
            "required": ["client", "affiliate", "insuranceObject", "insuranceRisks"],
            "properties": {
                "status": {
                "type": "string",
                "enum": ["CREATED", "CONFIRMED", "COMPLETED"]
                },
                "client": {
                "type": "string",
                "format": "ObjectId"
                },
                "affiliate": {
                "type": "string",
                "format": "ObjectId"
                },
                "insuranceObject": {
                "type": "string",
                "format": "ObjectId"
                },
                "insuranceRisks": {
                "type": "array",
                "items": {
                    "type": "string",
                    "format": "ObjectId"
                }
                }
            }
        },
...
        "Affiliate": {
            "type": "object",
            "required": [
                "name",
                "address"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "Agent": {
            "type": "object",
            "required": [
                "first_name",
                "last_name",
                "email",
                "birthdate",
                "password",
                "re_password",
                "tarrif_rate",
                "salary",
                "affiliate_name"
            ],
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "format": "email"
                },
                "birthdate": {
                    "type": "string",
                    "format": "date"
                },
                "password": {
                    "type": "string",
                    "format": "password"
                },
                "re_password": {
                    "type": "string",
                    "format": "password"
                },
                "tarrif_rate": {
                    "type": "number"
                },
                "salary": {
                    "type": "number"
                },
                "affiliate_name": {
                    "type": "string"
                }
            }
        },
...
        "InsuranceObject": {
        "type": "object",
        "required": ["insuranceType", "name", "description"],
        "properties": {
            "insuranceType": {
            "type": "string",
            "enum": ["MEDICAL", "AUTO", "TRAVEL", "BUSINESS"]
            },
            "name": {
            "type": "string"
            },
            "description": {
            "type": "string"
            }
        }
        },
        "InsuranceRisk": {
        "type": "object",
        "required": ["insuranceObject", "name"],
        "properties": {
            "insuranceObject": {
            "type": "string",
            "format": "ObjectId"
            },
            "name": {
            "type": "string"
            }
        }
    }
    }
}